requirements.txt
aiomysql
fastapi~=0.120.1
PyMySQL~=1.1.2
pydantic~=2.12.3
loguru~=0.7.3
uvicorn
aiohttp~=3.13.1
botocore~=1.40.49
jinja2~=3.1.6
prometheus-client

prometheus_client~=0.23.1
python-dotenv~=1.2.1
typing_extensions~=4.15.0
starlette~=0.48.0
prometheus_client~=0.23.1
GitPython~=3.1.45
typing_extensions~=4.15.0
markdown-it-py
atomicwrites
pymongo
